Update workflow user to 'youngsecurity'
Adjusted the GitHub Actions workflow configuration to use the 'youngsecurity' account for operations. This reflects the transition to a new account that's aligned with our current security protocols. No functional changes to the workflow are expected apart from this account change. Be sure to update secrets.USER_TIMEZONE if the new account has a different timezone requirement.